backend/requirements.txt
# Backend Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.19.0
# asyncpg removed (PostgreSQL driver not used; avoids Windows build tools on py3.13)
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0

# AI & External Services
openai==1.3.5
weasyprint==61.2
stripe==6.6.0

# Internationalization & Utilities
fastapi-babel==0.0.9
fastapi-cors==0.0.6
python-dotenv==1.0.0
httpx==0.26.0

# Advanced Features
websockets==12.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.2
